actorDAO updated

+++ src/main/java/com/jiangren/movie/controller/ActorController.java
@@ -20,25 +20,29 @@ public class ActorController {
     @GetMapping
     public ResponseEntity<List<ActorGetDto>> find(@RequestParam(required = false) String firstname,
                                                   @RequestParam(required = false) String lastname) {
-        if (firstname != null || lastname != null) {
-            return ResponseEntity.ok(actorService.getActorsByFirstnameOrLastname(firstname, lastname));
-        }
+        if (firstname != null || lastname != null) return ResponseEntity.ok(actorService.getActorsByFirstnameOrLastname(firstname, lastname));
         return ResponseEntity.ok(actorService.getAllActors());
     }
 
     @PostMapping
-    public ResponseEntity<ActorGetDto> create(@RequestBody ActorPostDto actorPostDto){
+    public ResponseEntity<ActorGetDto> create(@RequestBody ActorPostDto actorPostDto) {
         return ResponseEntity.ok(actorService.create(actorPostDto));
     }
 
     @PutMapping
-    public ResponseEntity<ActorGetDto> update(@RequestParam Long id, @RequestBody ActorPutDto actorPutDto){
-        return ResponseEntity.ok(actorService.update(id,actorPutDto));
+    public ResponseEntity<?> update(@RequestParam Long id, @RequestBody ActorPutDto actorPutDto) {
+        if (actorService.getById(id).size() == 0) return ResponseEntity.status(404).body("Id not found");
+        return ResponseEntity.ok(actorService.update(id, actorPutDto));
     }
 
     @DeleteMapping
-    public ResponseEntity<?> delete(@RequestParam Long id){
-        actorService.delete(id);
-        return ResponseEntity.ok("Deleted");
+    public ResponseEntity<?> delete(@RequestParam Long id) {
+        if (actorService.getById(id).size() == 0) return ResponseEntity.status(404).body("Id not found");
+        try {
+            actorService.delete(id);
+            return ResponseEntity.ok().build();
+        } catch (Exception e) {
+            return ResponseEntity.status(422).body("Foreign key constraint");
+        }
     }
 }
